+++ packages/article-skeleton/src/article-skeleton.js
@@ -11,7 +11,9 @@ import {
   WelcomeBanner,
   ArticleSidebar,
   UpdateButtonWithDelay,
-  Banner
+  Banner,
+  SocialEmbedsProvider,
+  useSocialEmbedsContext
 } from "@times-components/ts-components";
 import { spacing } from "@times-components/ts-styleguide";
 import UserState from "@times-components/user-state";
@@ -44,7 +46,6 @@ import Head from "./head";
 import PaywallPortal from "./paywall-portal";
 import StickySaveAndShareBar from "./sticky-save-and-share-bar";
 import insertDropcapIntoAST from "./contentModifiers/dropcap-util";
-import insertNewsletterPuff from "./contentModifiers/newsletter-puff";
 import insertInlineAd from "./contentModifiers/inline-ad";
 import { getIsLiveOrBreakingFlag } from "./data-helper";
 import setExternalLinkTargets from "./contentModifiers/setExternalLinkTargets";
@@ -93,6 +94,22 @@ const ArticleSkeleton = ({
   } = article;
 
   const [showVerifyEmailBanner, setShowEmailVerifyBanner] = useState(false);
+
+  const { isSocialEmbedAllowed, isAllowedOnce } = useSocialEmbedsContext();
+
+  useEffect(
+    () => {
+      // Trigger Twitter embed load when isSocialEmbedAllowed or isAllowedOnce switches to true
+      if (
+        (isSocialEmbedAllowed.twitter || isAllowedOnce.twitter) &&
+        window.twttr &&
+        window.twttr.widgets
+      ) {
+        window.twttr.widgets.load();
+      }
+    },
+    [isSocialEmbedAllowed.twitter, isAllowedOnce.twitter]
+  );
 
   const sidebarRef = useRef();
 
@@ -153,7 +170,6 @@ const ArticleSkeleton = ({
 
   const articleContentReducers = [
     insertDropcapIntoAST(template, dropcapsDisabled),
-    insertNewsletterPuff(section, isPreview, expirableFlags),
     insertInlineAd(isPreview),
     tagLastParagraph,
     setExternalLinkTargets
@@ -466,6 +482,12 @@ const ArticleSkeleton = ({
   );
 };
 
+const ArticleSkeletonWithProvider = props => (
+  <SocialEmbedsProvider>
+    <ArticleSkeleton {...props} />
+  </SocialEmbedsProvider>
+);
+
 ArticleSkeleton.propTypes = {
   ...articleSkeletonPropTypes,
   paidContentClassName: PropTypes.string
@@ -476,4 +498,6 @@ export { KeylineItem, ArticleKeylineItem } from "./keylines";
 
 export { ArticleLink };
 
-export default articleTrackingContext(withTrackScrollDepth(ArticleSkeleton));
+export default articleTrackingContext(
+  withTrackScrollDepth(ArticleSkeletonWithProvider)
+);