+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.93.3](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.93.2...@times-components/article-in-depth@3.93.3) (2024-12-12)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.93.2](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.93.1...@times-components/article-in-depth@3.93.2) (2024-12-11)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.93.1](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.93.0...@times-components/article-in-depth@3.93.1) (2024-12-10)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+# [3.93.0](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.92.6...@times-components/article-in-depth@3.93.0) (2024-12-05)
+
+
+### Features
+
+* **TMC-0000:** Implement social embeds ([#3989](https://github.com/newsuk/times-components/issues/3989)) ([4a4e7cd](https://github.com/newsuk/times-components/commit/4a4e7cdceec5fe7402b4403170013f325eec8a84))
+
+
+
+
+
 ## [3.92.6](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.92.5...@times-components/article-in-depth@3.92.6) (2024-12-03)
 
 **Note:** Version bump only for package @times-components/article-in-depth