+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.142.3](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.142.2...@times-components/article-skeleton@1.142.3) (2024-12-12)
+
+
+### Bug Fixes
+
+* **TMRS-575:** iframe component heights ([#4006](https://github.com/newsuk/times-components/issues/4006)) ([f4c3691](https://github.com/newsuk/times-components/commit/f4c36910a5a5d7e82dcfd10acfa62e3b3278154f))
+
+
+
+
+
+## [1.142.2](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.142.1...@times-components/article-skeleton@1.142.2) (2024-12-11)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.142.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.142.0...@times-components/article-skeleton@1.142.1) (2024-12-10)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+# [1.142.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.141.1...@times-components/article-skeleton@1.142.0) (2024-12-05)
+
+
+### Features
+
+* **TMC-0000:** Implement social embeds ([#3989](https://github.com/newsuk/times-components/issues/3989)) ([4a4e7cd](https://github.com/newsuk/times-components/commit/4a4e7cdceec5fe7402b4403170013f325eec8a84))
+
+
+
+
+
 ## [1.141.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.141.0...@times-components/article-skeleton@1.141.1) (2024-12-03)
 
 **Note:** Version bump only for package @times-components/article-skeleton